+++ tests/test_live_client.py
@@ -14,12 +14,12 @@ from unittest.mock import MagicMock
 import databento_dbn
 import pytest
 import zstandard
+from databento.common.constants import ALL_SYMBOLS
 from databento.common.cram import BUCKET_ID_LENGTH
 from databento.common.data import SCHEMA_STRUCT_MAP
 from databento.common.dbnstore import DBNStore
 from databento.common.error import BentoError
 from databento.common.publishers import Dataset
-from databento.common.symbology import ALL_SYMBOLS
 from databento.live import DBNRecord
 from databento.live import client
 from databento.live import gateway