+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 0.22.2 - TBD
+## 0.23.0 - TBD
+
+#### Enhancements
+- Added `map_symbols_csv` function to the `databento` module for using `symbology.json` files to map a symbol column onto a CSV file
+- Added `map_symbols_json` function to the `databento` module for using `symbology.json` files to add a symbol key to a file of JSON records
 
 #### Bug fixes
 - Fixed issue where a large unreadable symbol subscription message could be sent